further refactor custom rule creation from search filter
